SERVER-43910 Remove usage of BSONObjBuilder append unsigned method from lockInfo command
(partially cherry picked from commit 990b2ef30d7b8fa4db6af6d79da80e3664df9c21)

+++ src/mongo/db/commands/lock_info.cpp
@@ -95,7 +95,7 @@ public:
                 // The client information
                 client->reportState(infoBuilder);
 
-                infoBuilder.append("opid", clientOpCtx->getOpID());
+                infoBuilder.append("opid", static_cast<int>(clientOpCtx->getOpID()));
                 LockerId lockerId = clientOpCtx->lockState()->getId();
                 lockToClientMap.insert({lockerId, infoBuilder.obj()});
             }
